[C] Break MobileMenu out into a layout component
