+++ components/global/index.tsx
@@ -3,3 +3,5 @@ export { default as Footer } from "./Footer";
 export { default as Header } from "./Header";
 export { default as LogoPlaceholder } from "./LogoPlaceholder";
 export { default as DrawerController } from "./DrawerController";
+export { default as ProviderBar } from "./ProviderBar";
+export { default as GlobalSearch } from "./GlobalSearch";
