+++ components/layout/MobileMenu/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./MobileMenu";
+export { List as MobileMenuList } from "./MobileMenu.styles";
